reader refactor
